Add OrderItem snapshot model and pricing logic
Implemented the OrderItem entity to represent a snapshot of product data at the time of purchase.

The model captures immutable product attributes (identifier, name, unit price) alongside the selected quantity, ensuring historical consistency even if the product catalog changes.

Added validation rules to prevent invalid quantities and introduced line total calculation logic based on unit price and quantity.

This structure prepares the order aggregate for accurate subtotal and total computations while preserving domain integrity.

+++ src/main/java/domain/CartItem.java
@@ -2,7 +2,7 @@ package domain;
 
 public class CartItem {
     private Product product;
-    private double quantity;
+    private int quantity;
 
     public Product getProduct() {
         return product;
@@ -12,11 +12,11 @@ public class CartItem {
         this.product = product;
     }
 
-    public double getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(double quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
 }
